create separated modules for each menu item
