+++ apps/settings/js/panels/root/storage_media.js
@@ -1,23 +1,23 @@
 /* global DeviceStorageHelper */
 /**
- * Links the root panel list item with AppStorage.
+ * Links the root panel list item with MediaStorage.
  */
 define(function(require) {
   'use strict';
 
   var SettingsListener = require('shared/settings_listener');
-  var AsyncStorage = require('shared/async_storage');
-  var AppStorage = require('modules/app_storage');
 
-  function Storage() {
+  function MediaStoragePanel() {
     this._elements = null;
     this._enabled = false;
-    this.appStorage = null;
     this.defaultMediaVolume = null;
     this.defaultVolumeState = 'available';
   }
 
-  Storage.prototype = {
+  MediaStoragePanel.prototype = {
+    init: function media_storage_init(elements) {
+      this._elements = elements;
+    },
     /**
      * The value indicates whether the module is responding. If it is false, the
      * UI stops reflecting the updates from the root panel context.
@@ -31,48 +31,20 @@ define(function(require) {
 
     set enabled(value) {
       this._enabled = value;
-      var umsSettingKey = 'ums.enabled';
       var defaultMediaVolumeKey = 'device.storage.writable.name';
 
       if (value) { //observe
-        // ums master switch on root panel
-        this._elements.umsEnabledCheckBox.addEventListener('change', this);
-
-        SettingsListener.observe(umsSettingKey, false,
-          this._umsSettingHandler.bind(this));
-
-        // app storage
-        AppStorage.storage.observe('freeSize',
-          this.updateAppFreeSpace.bind(this));
-        this.updateAppFreeSpace();
-
         // media storage
         // Show default media volume state on root panel
         SettingsListener.observe(defaultMediaVolumeKey, 'sdcard',
           this._mediaVolumeChangeHandler.bind(this));
-
         window.addEventListener('localized', this);
       } else { //unobserve
-        this._elements.umsEnabledCheckBox.removeEventListener('change', this);
-
-        SettingsListener.unobserve(umsSettingKey,
-          this._umsSettingHandler.bind(this));
-
-        // app storage
-        AppStorage.storage.unobserve('freeSize',
-          this.updateAppFreeSpace.bind(this));
-
         // media storage
         SettingsListener.unobserve(defaultMediaVolumeKey,
           this._mediaVolumeChangeHandler.bind(this));
-
         window.removeEventListener('localized', this);
       }
-    },
-
-    _umsSettingHandler: function storage_umsSettingHandler(enabled) {
-      this._elements.umsEnabledCheckBox.checked = enabled;
-      this.updateUmsDesc();
     },
 
     _mediaVolumeChangeHandler:
@@ -85,83 +57,19 @@ define(function(require) {
       this.updateMediaStorageInfo();
     },
 
-    init: function storage_init(elements) {
-      this._elements = elements;
-    },
-
     handleEvent: function storage_handleEvent(evt) {
       switch (evt.type) {
         case 'localized':
-          this.updateAppFreeSpace();
           this.updateMediaStorageInfo();
           break;
         case 'change':
-          if (evt.target === this._elements.umsEnabledCheckBox) {
-            this.umsMasterSettingChanged(evt);
-          } else {
+          if (evt.target !== this._elements.umsEnabledCheckBox) {
             // we are handling storage state changes
             // possible state: available, unavailable, shared
             this.updateMediaStorageInfo();
           }
           break;
       }
-    },
-
-    // ums description
-    updateUmsDesc: function storage_updateUmsDesc() {
-      var localize = navigator.mozL10n.localize;
-      var key;
-      if (this._elements.umsEnabledCheckBox.checked) {
-        //TODO list all enabled volume name
-        key = 'enabled';
-      } else if (this.defaultVolumeState === 'shared') {
-        key = 'umsUnplugToDisable';
-      } else {
-        key = 'disabled';
-      }
-      localize(this._elements.umsEnabledInfoBlock, key);
-    },
-
-    umsMasterSettingChanged: function storage_umsMasterSettingChanged(evt) {
-      var checkbox = evt.target;
-      var cset = {};
-      var umsSettingKey = 'ums.enabled';
-      var warningKey = 'ums-turn-on-warning';
-      if (checkbox.checked) {
-        AsyncStorage.getItem(warningKey, function(showed) {
-          if (!showed) {
-            this._elements.umsWarningDialog.hidden = false;
-
-            this._elements.umsConfirmButton.onclick = function() {
-              cset[umsSettingKey] = true;
-              Settings.mozSettings.createLock().set(cset);
-
-              AsyncStorage.setItem(warningKey, true);
-              this._elements.umsWarningDialog.hidden = true;
-            }.bind(this);
-
-            this._elements.umsCancelButton.onclick = function() {
-              cset[umsSettingKey] = false;
-              Settings.mozSettings.createLock().set(cset);
-
-              checkbox.checked = false;
-              this._elements.umsWarningDialog.hidden = true;
-            }.bind(this);
-          } else {
-            cset[umsSettingKey] = true;
-            Settings.mozSettings.createLock().set(cset);
-          }
-        }.bind(this));
-      } else {
-        cset[umsSettingKey] = false;
-        Settings.mozSettings.createLock().set(cset);
-      }
-    },
-
-    // Application Storage
-    updateAppFreeSpace: function storage_updateAppFreeSpace() {
-      DeviceStorageHelper.showFormatedSize(this._elements.appStorageDesc,
-        'availableSize', AppStorage.storage.freeSize);
     },
 
     // Media Storage
@@ -243,7 +151,7 @@ define(function(require) {
     }
   };
 
-  return function ctor_root() {
-    return new Storage();
+  return function ctor_media_storage_panel() {
+    return new MediaStoragePanel();
   };
 });
